Uploading at 2am
This afternoon I got interested and began to read the Oracle Java JVM and .class File specification. And so to help me wrap my head around all of the details of the .class file format, I began to write a visualizer for decompiling the .class file format. After wayyy too many hours when I have other things I need to do, I now have about 600 lines of code and just want to go to sleep, so I have no idea how broken it is at this stage, but I'm just gonna commit this and go to bed. Hopefully I'll fix it all later, or maybe not.